LayoutTests/http/tests/security/video-cross-origin-caching-expected.txt


This test passes if you do not see a CORS error.

EVENT(playing)
EVENT(playing)
END OF TEST
